moving logic from controller into service layer
