+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/services/UIMetadataImportService.java
@@ -1,223 +1,242 @@
 package ca.corefacility.bioinformatics.irida.ria.web.services;
 
+import java.io.ByteArrayInputStream;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 import java.util.*;
 
-import org.apache.commons.csv.CSVFormat;
-import org.apache.commons.csv.CSVParser;
-import org.apache.commons.csv.CSVRecord;
-import org.apache.poi.hssf.usermodel.HSSFWorkbook;
-import org.apache.poi.ss.usermodel.*;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
+import javax.servlet.http.HttpSession;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Component;
+import org.springframework.web.multipart.MultipartFile;
 
 import ca.corefacility.bioinformatics.irida.exceptions.EntityNotFoundException;
 import ca.corefacility.bioinformatics.irida.exceptions.MetadataImportFileTypeNotSupportedError;
 import ca.corefacility.bioinformatics.irida.model.project.Project;
+import ca.corefacility.bioinformatics.irida.model.sample.MetadataTemplateField;
+import ca.corefacility.bioinformatics.irida.model.sample.Sample;
+import ca.corefacility.bioinformatics.irida.model.sample.metadata.MetadataEntry;
 import ca.corefacility.bioinformatics.irida.ria.utilities.SampleMetadataStorage;
+import ca.corefacility.bioinformatics.irida.ria.web.projects.ProjectControllerUtils;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
+import ca.corefacility.bioinformatics.irida.service.sample.MetadataTemplateService;
 import ca.corefacility.bioinformatics.irida.service.sample.SampleService;
 
-import com.google.common.base.Strings;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
+import com.google.common.io.Files;
 
 /**
- * UI service to handle parsing metadata files so they can be saved to the session.
+ * UI service to handle importing metadata files so they can be saved to the session.
  */
 @Component
 public class UIMetadataImportService {
 
 	private static final Logger logger = LoggerFactory.getLogger(UIMetadataImportService.class);
-
+	private final MessageSource messageSource;
 	private final ProjectService projectService;
 	private final SampleService sampleService;
+	private final MetadataTemplateService metadataTemplateService;
+	private final ProjectControllerUtils projectControllerUtils;
+	private final UIMetadataFileImportService metadataFileImportService;
 
 	@Autowired
-	public UIMetadataImportService(ProjectService projectService, SampleService sampleService) {
+	public UIMetadataImportService(MessageSource messageSource, ProjectService projectService,
+			SampleService sampleService, MetadataTemplateService metadataTemplateService,
+			ProjectControllerUtils projectControllerUtils, UIMetadataFileImportService metadataFileImportService) {
+		this.messageSource = messageSource;
 		this.projectService = projectService;
 		this.sampleService = sampleService;
+		this.metadataTemplateService = metadataTemplateService;
+		this.projectControllerUtils = projectControllerUtils;
+		this.metadataFileImportService = metadataFileImportService;
 	}
 
 	/**
-	 * Parse metadata from an csv file.
+	 * Upload CSV or Excel file containing sample metadata and extract the headers.  The file is stored in the session until
+	 * the column that corresponds to a {@link Sample} identifier has been sent.
 	 *
-	 * @param projectId   {@link Long} The project identifier.
-	 * @param inputStream The inputStream of the csv file.
-	 * @return {@link SampleMetadataStorage} contains the metadata from file.
-	 * @throws IOException thrown if the extension does not exist.
+	 * @param session   {@link HttpSession}
+	 * @param projectId {@link Long} identifier for the current {@link Project}
+	 * @param file      {@link MultipartFile} The csv or excel file containing the metadata.
+	 * @return {@link Map} of headers and rows from the csv or excel file for the user to select the header corresponding the
+	 * {@link Sample} identifier.
 	 */
-	public SampleMetadataStorage parseCSV(Long projectId, InputStream inputStream) throws IOException {
+	public SampleMetadataStorage createProjectSampleMetadata(HttpSession session, long projectId, MultipartFile file) {
+		// We want to return a list of the table headers back to the UI.
 		SampleMetadataStorage storage = new SampleMetadataStorage();
+		try {
+			String filename = file.getOriginalFilename();
+			byte[] byteArr = file.getBytes();
+			InputStream inputStream = new ByteArrayInputStream(byteArr);
 
-		CSVParser parser = CSVParser.parse(inputStream, StandardCharsets.UTF_8,
-				CSVFormat.RFC4180.withFirstRecordAsHeader()
-						.withTrim()
-						.withIgnoreEmptyLines());
-		List<Map<String, String>> rows = new ArrayList<>();
+			String extension = Files.getFileExtension(filename);
 
-		// save headers
-		Map<String, Integer> headers_set = parser.getHeaderMap();
-		List<String> headers_list = new ArrayList<>(headers_set.keySet());
-		storage.saveHeaders(headers_list);
-
-		// save data
-		for (CSVRecord row : parser) {
-			Map<String, String> rowMap = new HashMap<>();
-			for (String key : row.toMap()
-					.keySet()) {
-				String value = row.toMap()
-						.get(key);
-				rowMap.put(key, value);
+			// Check the file type
+			switch (extension) {
+			case "csv":
+				storage = metadataFileImportService.parseCSV(projectId, inputStream);
+				break;
+			case "xlsx":
+			case "xls":
+				storage = metadataFileImportService.parseExcel(projectId, inputStream, extension);
+				break;
+			default:
+				// Should never reach here as the uploader limits to .csv, .xlsx and .xlx files.
+				throw new MetadataImportFileTypeNotSupportedError(extension);
 			}
-			rows.add(rowMap);
+			inputStream.close();
+		} catch (FileNotFoundException e) {
+			logger.debug("No file found for uploading an excel file of metadata.");
+		} catch (IOException e) {
+			logger.error("Error opening file" + file.getOriginalFilename());
 		}
 
-		storage.saveRows(rows);
-		storage.setSampleNameColumn(findColumnName(projectId, rows.get(0)));
-		parser.close();
-
+		session.setAttribute("pm-" + projectId, storage);
 		return storage;
 	}
 
 	/**
-	 * Parse metadata from an excel file.
+	 * Add the metadata to specific {@link Sample} based on the selected column to correspond to the {@link Sample} id.
 	 *
-	 * @param projectId   {@link Long} The project identifier.
-	 * @param inputStream The inputStream of the excel file.
-	 * @param extension   The extension of the excel file.
-	 * @return {@link SampleMetadataStorage} contains the metadata from file.
-	 * @throws IOException thrown if the extension does not exist.
+	 * @param session          {@link HttpSession}.
+	 * @param projectId        {@link Long} identifier for the current {@link Project}.
+	 * @param sampleNameColumn {@link String} the header to used to represent the {@link Sample} identifier.
+	 * @return {@link Map} containing a complete message.
 	 */
-	public SampleMetadataStorage parseExcel(Long projectId, InputStream inputStream, String extension)
-			throws IOException {
-		SampleMetadataStorage storage = new SampleMetadataStorage();
-		Workbook workbook = null;
+	public Map<String, Object> setProjectSampleMetadataSampleId(HttpSession session, long projectId,
+			String sampleNameColumn) {
+		// Attempt to get the metadata from the sessions
+		SampleMetadataStorage stored = (SampleMetadataStorage) session.getAttribute("pm-" + projectId);
 
-		// Check the type of workbook
-		switch (extension) {
-		case "xlsx":
-			workbook = new XSSFWorkbook(inputStream);
-			break;
-		case "xls":
-			workbook = new HSSFWorkbook(inputStream);
-			break;
-		default:
-			// Should never reach here as the uploader limits to .csv, .xlsx and .xlx files.
-			throw new MetadataImportFileTypeNotSupportedError(extension);
+		if (stored != null) {
+			stored.setSampleNameColumn(sampleNameColumn);
+			Project project = projectService.read(projectId);
+			List<Map<String, String>> rows = stored.getRows();
+
+			// Remove 'rows' since they are now going to be sorted into found and not found.
+			stored.removeRows();
+			List<Map<String, String>> found = new ArrayList<>();
+			List<Map<String, String>> missing = new ArrayList<>();
+
+			// Get the metadata out of the table.
+			for (Map<String, String> row : rows) {
+				try {
+					// If this throws an error than the sample does not exist.
+					sampleService.getSampleBySampleName(project, row.get(sampleNameColumn));
+					found.add(row);
+				} catch (EntityNotFoundException e) {
+					missing.add(row);
+				}
+			}
+
+			stored.saveFound(found);
+			stored.saveMissing(missing);
 		}
 
-		// Only look at the first sheet in the workbook as this should be the file we want.
-		Sheet sheet = workbook.getSheetAt(0);
-		Iterator<Row> rowIterator = sheet.iterator();
+		return ImmutableMap.of("result", "complete");
+	}
 
-		List<String> headers = getWorkbookHeaders(rowIterator.next());
-		storage.saveHeaders(headers);
+	/**
+	 * Save uploaded metadata to the
+	 *
+	 * @param locale    {@link Locale} of the current user.
+	 * @param session   {@link HttpSession}
+	 * @param projectId {@link Long} identifier for the current project
+	 * @return {@link Map} of potential errors.
+	 */
+	public Map<String, Object> saveProjectSampleMetadata(Locale locale, HttpSession session, long projectId) {
+		List<String> DEFAULT_HEADERS = ImmutableList.of("Sample Id", "ID", "Modified Date", "Modified On",
+				"Created Date", "Created On", "Coverage", "Project ID");
+		Map<String, Object> errors = new HashMap<>();
+		Project project = projectService.read(projectId);
 
-		// Get the metadata out of the table.
-		List<Map<String, String>> rows = new ArrayList<>();
-		while (rowIterator.hasNext()) {
-			Map<String, String> rowMap = new HashMap<>();
-			Row row = rowIterator.next();
-			Iterator<Cell> cellIterator = row.cellIterator();
-			while (cellIterator.hasNext()) {
-				Cell cell = cellIterator.next();
+		SampleMetadataStorage stored = (SampleMetadataStorage) session.getAttribute("pm-" + projectId);
+		if (stored == null) {
+			errors.put("stored-error", true);
+		}
 
-				int columnIndex = cell.getColumnIndex();
-				if (columnIndex < headers.size()) {
-					String header = headers.get(columnIndex);
+		List<Sample> samplesToUpdate = new ArrayList<>();
 
-					if (!Strings.isNullOrEmpty(header)) {
-						// Need to ignore empty headers.
-						if (cell.getCellTypeEnum()
-								.equals(CellType.NUMERIC)) {
-								/*
-								This is a special handler for number cells.  It was requested that numbers
-								keep their formatting from their excel files.  E.g. 2.222222 with formatting
-								for 2 decimal places will be saved as 2.22.
-								 */
-							DataFormatter formatter = new DataFormatter();
-							String value = formatter.formatCellValue(cell);
-							rowMap.put(header, value);
-						} else {
-							cell.setCellType(CellType.STRING);
-							rowMap.put(header, cell.getStringCellValue());
+		List<Map<String, String>> found = stored.getFound();
+		if (found != null) {
+			// Lets try to get a sample
+			String sampleNameColumn = stored.getSampleNameColumn();
+			List<String> errorList = new ArrayList<>();
+			try {
+				for (Map<String, String> row : found) {
+
+					String name = row.get(sampleNameColumn);
+					Sample sample = sampleService.getSampleBySampleName(project, name);
+					row.remove(sampleNameColumn);
+
+					Map<MetadataTemplateField, MetadataEntry> newData = new HashMap<>();
+
+					Set<MetadataEntry> metadataEntrySet = new HashSet<>();
+
+					// Need to overwrite duplicate keys
+					for (Map.Entry<String, String> entry : row.entrySet()) {
+						// Make sure we are not saving non-metadata items.
+						if (!DEFAULT_HEADERS.contains(entry.getKey())) {
+							MetadataTemplateField key = metadataTemplateService.readMetadataFieldByLabel(
+									entry.getKey());
+
+							if (key == null) {
+								key = metadataTemplateService.saveMetadataField(
+										new MetadataTemplateField(entry.getKey(), "text"));
+							}
+
+							metadataEntrySet.add(new MetadataEntry(entry.getValue(), "text", key));
 						}
 					}
-				}
-			}
-			rows.add(rowMap);
-		}
-		storage.saveRows(rows);
-		storage.setSampleNameColumn(findColumnName(projectId, rows.get(0)));
 
-		return storage;
+					// Save metadata back to the sample
+					sampleService.mergeSampleMetadata(sample, metadataEntrySet);
+				}
+
+			} catch (EntityNotFoundException e) {
+				// This really should not happen, but hey, you never know!
+				errorList.add(messageSource.getMessage("metadata.results.save.sample-not-found",
+						new Object[] { e.getMessage() }, locale));
+			}
+
+			if (errorList.size() > 0) {
+				errors.put("save-errors", errorList);
+			}
+		} else {
+			errors.put("found-error",
+					messageSource.getMessage("metadata.results.save.found-error", new Object[] {}, locale));
+		}
+		if (errors.size() == 0) {
+			return ImmutableMap.of("success",
+					messageSource.getMessage("metadata.results.save.success", new Object[] { found.size() }, locale));
+		}
+		return errors;
 	}
 
 	/**
-	 * Extract the headers from an excel file.
+	 * Clear any uploaded sample metadata stored into the session.
 	 *
-	 * @param row {@link Row} First row from the excel file.
-	 * @return {@link List} of {@link String} header values.
+	 * @param session   {@link HttpSession}
+	 * @param projectId identifier for the {@link Project} currently uploaded metadata to.
 	 */
-	private List<String> getWorkbookHeaders(Row row) {
-		// We want to return a list of the table headers back to the UI.
-		List<String> headers = new ArrayList<>();
-
-		// Get the column headers
-		Iterator<Cell> headerIterator = row.cellIterator();
-		while (headerIterator.hasNext()) {
-			Cell headerCell = headerIterator.next();
-			CellType cellType = headerCell.getCellTypeEnum();
-
-			String headerValue;
-			if (cellType.equals(CellType.STRING)) {
-				headerValue = headerCell.getStringCellValue()
-						.trim();
-			} else {
-				headerValue = String.valueOf(headerCell.getNumericCellValue())
-						.trim();
-			}
-
-			// Leave empty headers for now, we will remove those columns later.
-			headers.add(headerValue);
-		}
-		return headers;
+	public void clearProjectSampleMetadata(HttpSession session, long projectId) {
+		session.removeAttribute("pm-" + projectId);
 	}
 
 	/**
-	 * Extract the headers from an excel file.
+	 * Get the currently stored metadata.
 	 *
-	 * @param projectId {@link Long} The project identifier.
-	 * @param row       {@link Row} First row from the excel file.
-	 * @return {@link String} column name.
+	 * @param session   {@link HttpSession}
+	 * @param projectId {@link Long} identifier for the current {@link Project}
+	 * @return the currently stored {@link SampleMetadataStorage}
 	 */
-	private String findColumnName(Long projectId, Map<String, String> row) {
-		String columnName = null;
-		Project project = projectService.read(projectId);
-		Iterator<Map.Entry<String, String>> iterator = row.entrySet()
-				.iterator();
-
-		while (iterator.hasNext() && columnName == null) {
-			String key = null;
-			String value = null;
-			try {
-				Map.Entry<String, String> entry = iterator.next();
-				key = entry.getKey();
-				value = entry.getValue();
-
-				if (sampleService.getSampleBySampleName(project, value) != null) {
-					columnName = key;
-				}
-			} catch (EntityNotFoundException entityNotFoundException) {
-				logger.trace("Sample " + value + " in project " + project.getId() + " is not found.",
-						entityNotFoundException);
-			}
-		}
-
-		return columnName;
+	public SampleMetadataStorage getProjectSampleMetadata(HttpSession session, long projectId) {
+		return (SampleMetadataStorage) session.getAttribute("pm-" + projectId);
 	}
 }
